requirements.txt
requests>=2.6.0
celery>=3.1.17
django>=1.7,<1.8
django-celery>=3.1.16
djangorestframework
markdown
requests_ntlm
gunicorn
dj-static
psycopg2
coverage==4.0a5
redis